requirements.txt
# APEX TRADER - Python Dependencies
# Engineer: Mane 🔥💙

# Binance API
python-binance==1.0.19

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Technical Analysis
TA-Lib==0.4.28
ta==0.11.0

# Web API & Dashboard
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0

# Logging
colorlog==6.8.0

